Move mappedConfig to separate file
